+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog 1.0.0].
 
 - prevent random CI failures due to missing Redis connections
 - GitHub action that allows manual cloning from master to staging with one-click for content testing
+- steps are listed in the order set in Contentful rather than database insert order
 
 ## [release-012] - 2021-05-11